Manually add `created_at` and `updated_at` to attributes
In Rails7 the timestamps column is automatically updated when insert_all and upsert_all are executed,
but before that it had to be updated manually.
https://github.com/rails/rails/pull/43003

There was a related pull request that presicion was not taken into account when auto-updating timestamps columns,
but this does not affect `solid_queue` table schema as there is no presicion setting.
https://github.com/rails/rails/pull/42993